+++ applications/PUfoam/Models/FoamConstruction/foamreconstr/src/nodes.hh
@@ -1,6 +1,6 @@
 #ifndef NODES_H
 #define NODES_H
 
-void makeNodeStruts(int ***, int, int, int, float **, int **);
+void makeNodeStruts(int ***, int, int, int, double **, int **);
 
 #endif
